requirements.txt
fastapi
uvicorn
cryptography
PyJWT
pytest
httpx
